# docs/swagger.yaml
definitions:
...
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/coolishbee/go-gin-sample/blob/main/LICENSE
  termsOfService: https://github.com/coolishbee/go-gin-sample
  title: Golang Gin API
  version: "1.0"
paths:
...
  /api/playerlist/{team_id}:
    get:
      parameters:
      - description: TEAM ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Get player list by team id
...
swagger: "2.0"